+++ api/src/constants/github.js
@@ -20,14 +20,6 @@
 
 'use strict';
 
-// TODO: move to constants folder
-export const ACCESS_CONTROL = {
-  SA_ROLE: 'devhub_kc_web',
-  REQUESTER_ROLE: 'devhub_kc_requester',
-  REVIEWER_ROLE: 'devhub_kc_reviewer',
-  NO_ROLE: 'unauthorized',
-};
-
 export const REALM_SCHEMA = {
   type: 'object',
   required: ['id', 'displayName', 'adminUser', 'idps', 'po'],
@@ -120,10 +112,4 @@ export const GITHUB_LABELS = {
   FAILED: 'request-failed',
   READY: 'ready-for-review',
   REJECTED: 'request-rejected',
-};
-
-export const KEYCLOAK_TERMS = {
-  BCEID: 'bceid',
-  IDIR: 'idir',
-  GITHUB: 'github',
 };
