+++ api/src/router/routes/github-webhook.js
@@ -25,7 +25,7 @@
 import { asyncMiddleware } from '@bcgov/common-nodejs-utils';
 import { Router } from 'express';
 import { validateSchema } from '../../libs/utils';
-import { PR_CONTENT_SCHEMA, GITHUB_LABELS } from '../../constants';
+import { PR_CONTENT_SCHEMA, GITHUB_LABELS } from '../../constants/github';
 import { setMailer } from '../../libs/email-utils';
 import { EMAIL_TYPE_TO_PATH, PR_ACTIONS } from '../../constants/email';
 import { EMAIL_TEST_CONTENT } from '../../constants/email-mock';
